requirements.txt
# Flask Web Framework
Flask>=2.3.0

# 核心数据处理库
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.3.0

# 数据库连接
psycopg2-binary>=2.9.0

# 配置文件处理
PyYAML>=6.0

# 科学计算
pandas>=1.3.0

# 网络请求（可选）
requests>=2.25.0

# 测试框架（可选）
pytest>=6.0.0
pytest-cov>=2.12.0

# 代码格式化（可选）
black>=21.0.0
flake8>=3.9.0 

gunicorn>=20.1.0